[tests] Updated FuseIterable specs
